+++ Editor/PopupWindows.cs
@@ -1,7 +1,8 @@
 ﻿using UnityEditor;
 using UnityEngine;
 
-namespace Saberslay.SlayTools.Editor.PopupWindows {
+namespace Saberslay.SlayTools.editor.PopupWindows
+{
     public class PopupWindows : EditorWindow {
         [MenuItem("Slayer Tools/Help/About")]
         public static void About() {
@@ -11,7 +12,7 @@ namespace Saberslay.SlayTools.Editor.PopupWindows {
             EditorGUILayout.BeginHorizontal();
             GUILayout.FlexibleSpace();
             GUILayout.Label("If you like to support me in development in this tool");
-            helper_functions.HelperFunctions.linkButton(110, 20, "Buy me a Coffee", "https://ko-fi.com/saberslay");
+            HelperFunctions.Button.linkButton(110, 20, "Buy me a Coffee", "https://ko-fi.com/saberslay");
             GUILayout.FlexibleSpace();
             EditorGUILayout.EndHorizontal();
         }