+++ CHANGELOG.md
@@ -7,8 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.0]
+## [1.0.0] - 16-04-2023
 
 ### Added
 - Create a parent gameobject for selected gameobjects
 - The ability to see the unity console in the game window in dev builds
+
+## [1.1.0] - 02-25-2023
+
+### Added
+- The ablity to capture a screen shots from the camra
+- Camra Prefab
+- The ablity to create files with a custom name
+- a timer function
+
+### Changed
+- Restructerd the helper functions